+++ src/NzbDrone.Core/MetadataSource/AuthorExtraMetadata.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+using NzbDrone.Core.Books;
+
+namespace NzbDrone.Core.MetadataSource
+{
+    public class AuthorExtraMetadata
+    {
+        public string ImageUrl { get; set; }
+        public string Overview { get; set; }
+        public List<Links> Links { get; set; }
+    }
+
+    public interface IAuthorExtraMetadataProvider
+    {
+        AuthorExtraMetadata GetAuthorExtraMetadata(string authorName);
+    }
+}